Minor tweaks on new.test.09 & sparse noise

+++ Spheres/Bonsai/new-test/Extensions/SphericalCoordinates_Sparsenoise.cs
@@ -11,7 +11,7 @@ using Bonsai.Reactive;
 [WorkflowElementCategory(ElementCategory.Transform)]
 public class SphericalCoordinates_Sparsenoise
 {
-    public IObservable<Tuple<int,Tuple<float,float,float>,Tuple<float,float,float,float,float>, Tuple<float,float>>> Process(IObservable<Tuple<Tuple<ElementIndex<Tuple<Tuple<double, double>, double>>, float>, float>> source)
+    public IObservable<Tuple<int,Tuple<float,float,float>,Tuple<float,float,float,float,float>, Tuple<float,float>>> Process(IObservable<Tuple<Tuple<ElementIndex<Tuple<Tuple<double, double>, double>>, float>, int>> source)
     {
         return source.Select(value =>{
             int SphereID = value.Item1.Item1.Index;
@@ -19,7 +19,7 @@ public class SphericalCoordinates_Sparsenoise
             float Elevation = Convert.ToSingle(value.Item1.Item1.Value.Item1.Item2);
             float Depth = Convert.ToSingle(value.Item1.Item1.Value.Item2);
             float OriginalSize = value.Item1.Item2;
-            float MouseZ = value.Item2;
+            float MouseZ = (float)value.Item2;
 
 
             // Rescale size of sphere
